+++ benchmarks/flask_sqli/scenario.py
@@ -1,22 +1,33 @@
 import bm
-import bm.flask_utils as flask_utils
-from utils import _post_response
+from bm.flask_utils import FlaskScenarioMixin
 
 
-class FlaskSQLi(bm.Scenario):
-    tracer_enabled = bm.var_bool()
-    profiler_enabled = bm.var_bool()
-    debugger_enabled = bm.var_bool()
-    appsec_enabled = bm.var_bool()
-    iast_enabled = bm.var_bool()
-    post_request = bm.var_bool()
-    telemetry_metrics_enabled = bm.var_bool()
-
+class FlaskSQLi(FlaskScenarioMixin, bm.Scenario):
     def run(self):
-        with flask_utils.server(self, custom_post_response=_post_response) as get_response:
+        app = self.create_app()
 
-            def _(loops):
-                for _ in range(500):
-                    get_response()
+        # Setup the request function
+        headers = {
+            "SERVER_PORT": "8000",
+            "REMOTE_ADDR": "127.0.0.1",
+            "HTTP_HOST": "localhost:8000",
+            "HTTP_ACCEPT": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,"
+            "image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9",
+            "HTTP_SEC_FETCH_DEST": "document",
+            "HTTP_ACCEPT_ENCODING": "gzip, deflate, br",
+            "HTTP_ACCEPT_LANGUAGE": "en-US,en;q=0.9",
+            "User-Agent": "dd-test-scanner-log",
+        }
 
-            yield _
+        def make_request(app):
+            client = app.test_client()
+            return client.post("/sqli", headers=headers, data={"username": "shaquille_oatmeal", "password": "123456"})
+
+        # Scenario loop function
+        def _(loops):
+            for _ in range(loops):
+                res = make_request(app)
+                assert res.status_code == 200
+                res.close()
+
+        yield _
